refactor: 필요없는 Repository 삭제
