--- src/main/java/com/tm/gogo/HikingTrailRepository.java
@@ -1,7 +1,0 @@
-package com.tm.gogo;
-
-import com.tm.gogo.domain.hiking_trail.HikingTrail;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HikingTrailRepository extends JpaRepository<HikingTrail, Long> {
-}
