--- src/main/java/com/tm/gogo/hikingLog/HikingLogImageRepository.java
@@ -1,6 +1,0 @@
-package com.tm.gogo.hikingLog;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HikingLogImageRepository extends JpaRepository<HikingLogImage, Long> {
-}
